Introduce friendly CLI interface, refactor
I'm not planning to make a GUI for this because that makes
cross-compiling much more difficult. It's also more room for issues.

A friendly, step-by-step, drag-and-drop command-line interface
should be sufficient for almost all users.

Additionally, a massive refactor under the hood to reduce the
amount of code in one file.