cpp: support replacing multiple occurrences
Fixes patches appearing to not apply (strings occur twice; once in
/sbin/recovery binary and once (at least) in twres strings)

Also fixes the actual patch having no effect, as it was trying to
replace 0 bytes with 0 bytes. Reason is because that patch starts with
a null byte, and by default literals are actually C strings that get
passed to the std::string constructor. Explicitly calling that
constructor with a length fixes this.